requirements.txt
aiogram==3.15.0
httpx==0.28.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
isodate==0.7.2
cachetools==5.3.2
lxml==5.1.0
